requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docx2txt==0.8
email_validator==2.2.0
faiss-cpu==1.10.0
fastapi==0.115.11
fastapi-cli==0.0.7
flake8==7.1.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
# Removed conflicting jwt package
# jwt==1.3.1
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.44
langchain-openai==0.3.8
langchain-text-splitters==0.3.6
langsmith==0.3.13
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mccabe==0.7.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.3
openai==1.66.3
orjson==3.10.15
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.3.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.1
PyJWT==2.10.1
pypdf==5.3.1
pytest==8.3.5
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
yarl==1.18.3
zstandard==0.23.0
